Replace CudaMatcher with TranslationRule class
The patch replaces CudaMatcher class with TranslationRule. The change is not
only about the renaming things, it's also introduces a new class hierarchy for
AST traversal rules with ASTTraversal base class. The idea is that any
descendant of ASTTraversal can be used to walk AST, but only TranslationRule has
ability to generate new replacements for the input program. It is expected that
in future the hierarchy will be extended with Analysis class that will not be
capable to produce replacements, but it will be possible to request result of
Analysis in another ASTTraversal.

Change-Id: Id036ee6a52ac517beb94ebbd7a320db6808611b6
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/138991
Reviewed-by: Bozhenov, Nikolai <nikolai.bozhenov@intel.com>
Tested-by: Bozhenov, Nikolai <nikolai.bozhenov@intel.com>