Add translation of kernel calls.
The function being called as a kernel is wrapped in a command group
lambda and a kernel lambda. The grid and thread dimensions are passed to
the constructor for nd_range<3>. In the function call, the nd_item<3>
object from the SYCL-kernel is forwarded. The command group is submitted
to the default queue of the current device through the compatibility
layer.

Change-Id: I86d70a552c0b103a016f5056de8943544f832efe
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/145572
Reviewed-by: German, Leonid <leonid.german@intel.com>
Tested-by: German, Leonid <leonid.german@intel.com>