Fix for CTST-124: Save dim3Ctor replacements, so they can be reused by overlapping rules (kernelCallExpr)
The translations for nested translation rules (such as the dim3
constructors and memberfield expressions) are saved in a Stmt->String map
(StmtStringMap).  When an overlapping rule are applied, these
translations can be retrived from the map instead of recomputed.  This
avoids code duplication.

Before the various transformations are converted into actual
replacements they sorted into groups, so that the getReplacement()
method can be called according to the nesting level; deeper nested
transformations are called first.

Also, a variant of getStmtSpelling, called getStmtSpellingWithTransforms
is introduced.  It will return the spelling of a given Stmt node with
the translations from the StmtStringMap applied.

The nd_range<1> and nd_range<2> prettification is undone as well, since
this is not legal SYCL code, as it turns out.

Change-Id: I7d2c1e1635572c505c42405bc8c224a04c7a463d
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/154215
Reviewed-by: Jensen, Peter <peter.jensen@intel.com>
Tested-by: Jensen, Peter <peter.jensen@intel.com>