Add a framework to fix non-type-template-paremter used in array size expression not replaced.
1. add a expression analysis framework, which can analysis expression
and get replacements of the expression if needed. The framework is
used while expression is matched, but more analysis is needed, or we
need the whole expression info, not only the replaced sub expression.

eg. If non-type-template-paremter is used in __shared__ array size
expression, we need to get the whole expression string with
non-type-template-paremter replaced.

2. fix issued that non-type-template-paremter used in array size
expression not replaced when the accessor is declared. The
non-type-template-paremter is still in range arguments, and now
it's replaced with corresponding template argument.

Change-Id: I9f26e86a9683f45a4f62c77b5e14deca2c4bbfaf
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/173374
Reviewed-by: Zhang Ziran <ziran.zhang@intel.com>
Tested-by: Zhang Ziran <ziran.zhang@intel.com>