Fix an issue that max/min functions were not migrated in windows
The header file “__clang_cuda_runtime_wrapper.h" includes "host_config.h”
with __CUDACC__ undefined. However, the contents of "host_config.h" is
protected by __CUDACC__, if __CUDACC__ is not defined, the contents of
"host_config.h" would not be included in “_clang_cuda_runtime_wrapper.h”.
So it results in that the macro NOMINMAX is not defined in "host_config.h",
and max/min functions are Macro, and will not be migrated. This behavior
is not consistent between clang and nvcc.

Refer to CTST-517.

Change-Id: I75ef758f6c8fe1fbf1a12699d25da4c1d858e134
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/188743
Reviewed-by: Wang, Zhiming <zhiming.wang@intel.com>
Tested-by: Wang, Zhiming <zhiming.wang@intel.com>