Fix an issue in migrating constant array size.
For following code:
“__device__ float tile[k_blockDimMaxY + 2*HALF_LENGTH][k_blockDimX+ 2*HALF_LENGTH];”
Before this patch, it is migrated to “syclct::device_memory<float, 2> tile(24, 80);”,
which 24/80 is hard coded.
With this patch, it is migrated to “syclct::device_memory<float, 2> tile
(k_blockDimMaxY + 2*HALF_LENGTH, k_blockDimX + 2*HALF_LENGTH);”.

But for device memory variables declared inside device functions, due to
the declaration contexts of migrated code and original code are different,
the constant variable and macro used in size expr may be undefined, so
in this case hard coded size is still used.

Ref to CTST-552.

Change-Id: Id543c30cb8a52f3fd3c0b3b37aaa1ec7afdd90b2
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/189406
Reviewed-by: Wang, Zhiming <zhiming.wang@intel.com>
Tested-by: Wang, Zhiming <zhiming.wang@intel.com>