Check if a stmt is inside a branch/loop W/O curly braces before removing the whole line
Before removing the whole line where a statement is replaced
with an empty string, further check whether the statement is
a single-line statement of a if/while/do while/for statement
without curly braces.

For example, in either case of the following, the lines "stmt;"
are guaranteed not to be removed.

if (...)
  stmt;

while (...)
  stmt;

do
  stmt;
while (...);

for (...)
  stmt;

Refer to CTST-665.

Change-Id: Iaa562cd94a0f3480b4fd035f9f65b1ac3c46f612
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/192891
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Reviewed-by: Zhenhao Tang <zhenhao.tang@intel.com>