Make sure math functions with std:: prefix will not be migrated
Previously, syclct depended on DeviceAttr to determine if a function
was a device function instead of a host function. However, clang
wrappers add __device__ attribute to std::min and std::max, making
our assumption inaccurate. As a result, the tool now treat all
functions with std:: prefix as host functions to make sure math
functions with std:: prefix will not be migrated.

Change-Id: I9d798964a25409f2fd4ee60a4a658125bf88fe2c
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/192898
Reviewed-by: Zhenhao Tang <zhenhao.tang@intel.com>
Tested-by: Zhenhao Tang <zhenhao.tang@intel.com>