Fix an issue that comments may break multiline macro definitions
Previously, when a comment is inserted to a multiline macro defintion
with backslashes, the defintion may be broken. For example,
"#define M\
  someCodeThatBringsMigrationWarnings"
was migrated to
"#define M\
  /*
    some warning messages
  */
  someCodeThatBringsMigrationWarnings"
which didn't have the same semantics as the original code and was
invalid in some cases.

This patch move the comments to the beginnings of the defintions,
resulting in
/*
  some warning messages
*/
"#define M\
  someCodeThatBringsMigrationWarnings"

Change-Id: I084b568d8fc0f8749caf39eedf610b7081847f13
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/194480
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Reviewed-by: Zhang Ziran <ziran.zhang@intel.com>
Reviewed-by: Wang, Zhiming <zhiming.wang@intel.com>