Fix a bug that a global variable is used without its extern declaration
E.g. A global variable is defined and referred to in a CUDA kernel function in demo.cu,
// demo.cu
volatile __device__ int g_mutex=0;
__global__ void Reset_kernel_parameters(void)
{
    g_mutex=0;
}

And the kernel function’s declaration is in demo.h,
// demo.h
__global__ void Reset_kernel_parameters(void);

When the host side file host.cpp calls the CUDA kernel function,
it ordinarily includes the header file demo.h,
// host.cpp
void test() {
  Reset_kernel_parameters<<<1,1>>>();
}

After migration, the migrated file of demo.cpp will access the global
variable explicitly, because an accessor is needed to have access to this
global variable from the kernel function.
So the advanced extern declaration for this global
variable is needed.

Refer to CTST-634.

Change-Id: Ifc1559487f8a3476a8782bb6e1ba9899925f5eff
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/200344
Reviewed-by: Chenwei Sun <chenwei.sun@intel.com>
Tested-by: Chenwei Sun <chenwei.sun@intel.com>