Fix a Klocwork issue on potential buffer overflow from unvalidated input
Klockwork complains that we direclty use the C-style string return from
unsafe external function 'getenv' without validation. This may brings
potential buffer overflow if the C-sytle string is not properly
null-teminated.

This patch validates the C-style string by setting a max-length threshold.
If the length of the C-sytle string is greater than or equal to INT_MAX,
we ignore this string and use a default string "." as the HOME
environment varaible.

Change-Id: Ib752e24412d7bac1e41f35e55acabb61288932cc
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/201395
Reviewed-by: Wang, Zheng W <zheng.w.wang@intel.com>
Reviewed-by: Zhang Ziran <ziran.zhang@intel.com>
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>