[CTST-1034] Fix the issue that "dpct -p=./", "dpct a.cu b.cu" don't work
"dpct -p=./" doesn't work before this patch, while it is expected to work.
"dpct a.cu b.cu" doesn't work before this patch, while it is expected to work.
If input source file doesn't exist, set --in-root to current folder.
If input source file exists in the commandline,
set --in-root to the directory of the first input source file.


Change-Id: Ie4e8de656db1cb35ee815ceb67f98be822e87aac
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/214824
Reviewed-by: Wang, Zhiming <zhiming.wang@intel.com>
Tested-by: Wang, Zhiming <zhiming.wang@intel.com>