[CTST-1048] Fixed the issue in the migration of func-like macro caused by wrong begin or end location calculation of macro.
E.g. For fun_call(..., MACRO_Sample(arg1, arg2),...)
When dpct uses getSpellingLoc to capture the range of MACRO_Sample(arg1, arg2),
the result differs according to the macro definition.

1. #define MACRO_Sample(arg1, arg2)  arg1 + arg2  --> both Begin/End at call loc
2. #define MACRO_Sample(arg1, arg2)  arg1 + arg2 + 1  --> Begin at call loc, End at def loc
3. #define MACRO_Sample(arg1, arg2)  1 + arg1 + arg2  --> Begin at def loc, End at call loc
4. #define MACRO_Sample(arg1, arg2)  1 + arg1 + arg2 + 1 --> both Begin&End at def loc

None of the results are correct, we need to use ExpansionLoc while
we want to capture the complete string of a function-like macro.

However, in other cases such as, MACRO_CHECK_RESULT(foo()), we
still need to use getSpellingLoc to get the correct location of foo().

Therefore, this patch uses the preprocessing result of an Expr to identify which AST API should be used.

This patch corrects the begin and end location calculation for arg1, arg2 and MACRO_Sample.

Change-Id: I1c8ace9549f569256e509b6dcddc6b994fc3b8a8
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/226914
Reviewed-by: Wang, Zhiming <zhiming.wang@intel.com>
Tested-by: Wang, Zhiming <zhiming.wang@intel.com>