[CTST-1296] Migrate function-call-parameter inside the command group
If a parameter of a kernel call is the a function call expression,
then a helper variable capturing the result of the function call
will be declared inside the command group and the variable will be
passed to the kernel.
This patch fixes an issue that the above expectation is not achieved
in several cases of templates.

Change-Id: I8d8bb11c05fd3c45d635ca002ab37f766ef4cd73
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/238261
Reviewed-by: Zhenhao Tang <zhenhao.tang@intel.com>
Tested-by: Zhenhao Tang <zhenhao.tang@intel.com>