[CTST-1246][Part 2] Fix crash when a file with shared var is migrated twice
Root cause:
ReplaceMap record all __shared__ variable decl.
The key of an entry in ReplaceMap is the RawEncoding of a SL.
The value of the entry is a ReplaceVarDecl.
ReplaceVarDecl will be destructed after getReplacement.
When migrating the same file again, DPCT will reuse the ReplaceMap,
and the RawEncoding of the same VarDecl will be the same which leads
DPCT reuse a destructed ReplaceVarDecl.

Solution:
Erase the entry from the ReplaceMap in getReplacement.

Change-Id: Ib7aabfaf235f0aa671842d15df5fc9d5318b5628
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/245304
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Reviewed-by: Wang, Zhiming <zhiming.wang@intel.com>