[CTST-1353] Remove unnecessary extern declaration
When processing a file, this patch records all header files included by 
current file. Then it can know whether the header file which contains 
the global variable definition is included or not by current file. If 
the definition is included, then there is no need to add extra "extern"
declaration for the global variable.


Change-Id: I9084f85699de62c4c1fba78b8022b51cb34580e2
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/250153
Reviewed-by: Wang, Zhiming <zhiming.wang@intel.com>
Tested-by: Wang, Zhiming <zhiming.wang@intel.com>