[CTST-1487][intercept-build]Remove duplicate file entry in compilation database.
The root cause is libear.so intercepts both the intermediate state command and
final state command for the same source file in some case.

For the intermediate state command, some options have not been analyzed finally,
such as '...,\"-O0\",\"-g\" ...', so invalid options maybe appeared in this kind of command.

While for the final command, all the options has been analyzed finally, there is so such
unanalyzed options in the command line.


So,for duplicate file in the compilation database, the patch keeps the final command
entry and remove the intermediate state one.


Change-Id: Ifc766693732b56bfa8d7b045cf9ee15e6d6a7808
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/259461
Reviewed-by: Wang, Zhiming <zhiming.wang@intel.com>
Tested-by: Wang, Zhiming <zhiming.wang@intel.com>