[CTST-1505] Fix the issue that --extra-arg cannot specify -x language option
-x<language> option makes the following files are treated as <language>
files until another -x option occurs.

Before this patch, dpct insert -x cuda at the begin of command;
The option in the extra-arg is inserted at the end of the command;
And the file name is between the above two options.
So the file is treated as a cuda file.

In this patch, when without database, if dpct detects there is
-x<language> in the extra-arg, dpct will insert this -x<language> at the
begin of the command and will not insert -xcuda.

Change-Id: Ie689b3e4ac4e5e879004507ba7c28c9f6765e128
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/259220
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Reviewed-by: Wang, Zhiming <zhiming.wang@intel.com>