[CTST-1501] Keep the indent of the original code
The formatted code use spaces as indent. The length is same as the original code.
The indent of the inserted warning is same as the indent of the next line of the warning.
If the original indent uses tab, tool treats 1 tab == 8 spaces to do formatting.
User can use .clang-format file to specify the TabWidth value.

Change-Id: Ia2ddc397679328b67713872cf8c2d7783f6acfe2
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/269520
Reviewed-by: Zhiwei, Jiang <zhiwei.jiang@intel.com>
Tested-by: Zhiwei, Jiang <zhiwei.jiang@intel.com>