[CTST-1690] Support multi level func-like macro in function rewriter
Original:
1 level func-like macro
#define FOO(x) sqrt(x)  -->  The rewriter will reuse 'x'

Issue:
Multi-level func-like macro
#define FOO(x) sqrt(x)
#define FOO2(x,y) powf(x,y)
FOO(FOO2(1,2)); --> The rewriter fails to use 'x','y' when rewrite FOO2

Solution:
Add a new candidate in getLocInCallSpelling with following steps
   (1) Remove outer func-like macro with getRangeInsideFuncLikeMacro()
   (2) Trace back all non-func-like macro with getTheOneBeforeLastImmediateExapansion()
   (3) Find the lowest common ancestor of Begin/End with getTheLastCompleteImmediateRange()

e.g. The matcher matches the first arg of powf, "1".
DPCT will call SM.getSpellingLoc twice in (1), and the immediate expansion
location will be "1" instead of "FOO" or "FOO2".
And in (2), DPCT will keep calling SM.getImmediateExpansionRange to get the
location of "x" in the definition of FOO2.
If the Begin/End location after (2) cannot form a complete range,
step (3) will find the common ancestor of the Begin/End in multi-level macro define.

Change-Id: I0cfaf7c5a8dd7c4709baaabee1689e91f34ccbfd
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/269843
Reviewed-by: Wang, Zhiming <zhiming.wang@intel.com>
Tested-by: Wang, Zhiming <zhiming.wang@intel.com>