[CTST-1989] Support passing func name in macro arg when no rewrite needed
Add NoRewriteFuncNameRewriter:
In which case, the callee name and the args will be migrated separately.
The arguments will be migrated in ExprAnalyze.
The Callee name will be migrated with the new name returned by NoRewriteFuncNameRewriter.rewrite().

Fix a small bug of deleting "__device__"

Move getCalleeName to Utility.h/cpp and expand to support cases in macro.

Change-Id: I70cf492c31c2936c196b9b5ae0d77822e19e7480
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/309188
Reviewed-by: Wang, Zhiming <zhiming.wang@intel.com>
Tested-by: Wang, Zhiming <zhiming.wang@intel.com>