[CTST-1995]Fix issue that test case Scan compilation failed with "no matching function for call to 'reduce'"
Root cause is that CUDA kernel in instantiation of template function is
matched firstly in EventAPICallRule::handleTargetCalls before template
function's prototype. In such cases the kernel call info are built
with substitution type instead of template type.

This patch make EventAPICallRule::handleTargetCalls not to process
CUDA kernel when  CUDA kernel is called in an instantiation of template function




Change-Id: Id408265fa8af18067eed5e71aedff25af952860e
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/309538
Reviewed-by: Wang, Zhiming <zhiming.wang@intel.com>
Tested-by: Wang, Zhiming <zhiming.wang@intel.com>