[CTST-2146] Fix __CUDA_ARCH__ migration on Windows caused by mixing using of lowercase and uppercase drive letter
The root cause of the incorrect migration is:
The path saved in ProcessedFile begins with a lowercase char
(e.g., d:/path/to/file.cu).
But paths saved in HostDeviceFDefIMap, HostDeviceFCallIMap and
HostDeviceFDeclIMap begin with an uppercase char (e.g., D:/path/to/file.cu).
Since the paths are different, the second parsing is skipped.

So this patch added a function to unify the drive letter to uppercase
char on Windows.

Change-Id: I9eaf399dc10052d6e55744b565656e9bc35ea043
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/327446
Reviewed-by: Wang, Zhiming <zhiming.wang@intel.com>
Tested-by: Wang, Zhiming <zhiming.wang@intel.com>