[CTST-2220] Fix the yaml loading issue when using --custom-helper-name option
Before:
When user specify --custom-helper-name=test, the feature info saved in
yaml is using test.hpp as file name.
When user do incremental migration, dpct will not load test.hpp file and
discard all the features used in test.hpp.
After:
If the file name is same as the MainHelperFileName saved in previous
yaml, treat this file as dpct.hpp

Change-Id: Ib4d97a5285838f85d42c66b48d7565e6a4ccab53
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/335899
Reviewed-by: Wang, Zhiming <zhiming.wang@intel.com>
Tested-by: Wang, Zhiming <zhiming.wang@intel.com>