[CTST-2240] Fix the issue that dpct::byte_level_permute has different behaviors on Linux and Windows platform
Root cause is that unsigned long is 32 bits or 64 bits which depends on Linux
and Window platform. 64 bits on Linux is an expected behavior, while 32 bits
on Window will make byte_level_permute failed in verification.

This patch uses uint64_t instead, whose size is always 64 bit both
on Windows and Linux platform.

Change-Id: Iee8869226efb6ec3e5a236917c9f4838b62033b0
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/335905
Reviewed-by: Zhiwei, Jiang <zhiwei.jiang@intel.com>
Reviewed-by: Chenwei Sun <chenwei.sun@intel.com>
Tested-by: Chenwei Sun <chenwei.sun@intel.com>