[SYCLomatic] Fix un-migrated issue caused by multi CUDA path specified by option --cuda-include-path and -I (#407)
Issue:
If two different CUDA header file paths are used in option --cuda-include-path
and option -I (or CPATH), the atomic functions are not migrated.

Analysis:
The tool use path specified by option --cuda-include-path to determine whether an API is from
CUDA or not. But in above case, the atomic function's declaration may be from
the path specified by option -I but not option --cuda-include-path, so the tool treat the API is non-CUDA API, and skip the migration.
Prepending another -I<cuda-include-path> before ToolInvocation.run() does not fix the problem. Because when generating cc1 args, the CudaInstallationDetector appends -internal-isystem=<detected path> in cc1 args, and the logic of removing duplicated path removes the value specified by -I<cuda-include-path>.

Fix:
1.	Prepend another -I<cuda-include-path> before ToolInvocation.run().
2.	In CudaInstallationDetector, don’t append -internal-isystem=<detected path> in cc1 args.

Signed-off-by: Jiang, Zhiwei <zhiwei.jiang@intel.com>