[SYCLomatic] Add block_tile_memory migration.
If the compute capability is higher than 8.0, the `block_tile_memory` function can be omitted. Therefore, in the migration code, the function is removed directly. As for the `this_thread_block` function, migrate it directly to the `experimental::this_sub_group()` free queries function, regardless of whether the function needs input parameters or not.

Signed-off-by: Chen, Sheng S <sheng.s.chen@intel.com>