[SYCLomatic] Add workaround for templated kernel call in macro (#1287)
clang preprocessor splits the ">>" to 2 ">"s in the template arg of a kernel call expr and the splited token is stored in "scratch space". However, when calling getImmediateRange at such location, the location is shifted right for 1 char if the kernel call expr is in a macro. An observation is that when this happens, both ">>"(greatergreater) and the later ">"(greater) are valid tokens. Which means the tokens overlap each other.

The WA detects the token overlaping of ">>" and ">", and then marking a flag for not skipping the last token when getting the end location of the typeloc in a kernelcall's template arg.

Signed-off-by: Huang, Andy <andy.huang@intel.com>