[SYCLomatic] Fix the cmath not found issue
On the ubuntu, when user installed the gcc package by "apt install
gcc-<version>" and not installed the libstdc++ package, the clang will
search the /usr/lib/gcc/x86_64-linux-gnu to get the latest GCC version
as default one. The PR will compare all the candidate and use the gcc
and libstdc++ all installed version.

Signed-off-by: Chen, Sheng S <sheng.s.chen@intel.com>